requirements.txt
requests==2.24.0
python-dotenv==0.14.0
boto3==1.16.20